When reading, read bytes instead of Strings
This is an API change, but since the bindings are fairly new I thought
we could still do this. Since reading bytes is also better.

+++ src/main/java/com/ceph/rados/jna/Rados.java
@@ -66,7 +66,7 @@ public interface Rados extends Library {
     int rados_write(Pointer ioctx, String oid, byte[] buf, int len, long off);
     int rados_write_full(Pointer ioctx, String oid, byte[] buf, int len);
     int rados_append(Pointer ioctx, String oid, byte[] buf, int len);
-    int rados_read(Pointer ioctx, String oid, byte[] buf, long len, long off);
+    int rados_read(Pointer ioctx, String oid, byte[] buf, int len, long off);
     int rados_remove(Pointer ioctx, String oid);
     int rados_trunc(Pointer ioctx, String oid, long size);
     int rados_clone_range(Pointer ioctx, String dst, long dst_off, String src, long src_off, long len);
